+++ CHANGELOG.md
@@ -4,9 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.2.0
 ### Changed
-- `parseServiceUrl` and `parseServerUrl` now return `null` if matches not found
-- `parseServiceUrl` deprecated, but delegates to `tryParseServiceUrl`,
-- `parseServiceUrl` deprecated, but delegates to `tryParseServerUrl`
+- `parseServiceUrl` and `parseServerUrl` now return the passed in url, stripped of any query string, if matches not found
 
 ## 1.0.1
 ### Fixed